requirements.txt
django>=6.0,<7.0
djangorestframework>=3.16,<3.17
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
python-decouple>=3.8,<4.0
whitenoise>=6.7,<7.0
gunicorn>=21.0,<23.0
